Remove duplicate code and restore functionality disabled during refactor.
