+++ Native/Messages/Abstractions/IMessageHandlerRegistrar.cs
@@ -6,6 +6,11 @@
         bool Register(IMessageHandler<TMessage> handler, 
             RegistrationSettings settings, params IIntercept<TMessage>[] interceptors);
 
+
+        bool Register(IMessageHandler<TMessage> handler,
+            params IIntercept<TMessage>[] interceptors) =>
+                Register(handler, default, interceptors);
+
         void Unregister(IMessageHandler<TMessage> handler);
     }
 }
