+++ Native/Messages/Abstractions/IMulticastMessageHandler.cs
@@ -1,6 +1,7 @@
-﻿namespace Chopsticks.Messages.Abstractions
-{
+﻿using Chopsticks.Messages.Abstractions;
 
+namespace Chopsticks.Messages
+{
     public interface IMulticastMessageHandler<TMessage> : 
         IMessageHandler<TMessage>
     {
