+++ Native/Messages/SyncHandlingPromiseSource.cs
@@ -1,0 +1,19 @@
+﻿using System;
+
+namespace Chopsticks.Messages
+{
+    public class SyncHandlingPromiseSource :
+        BaseHandlingPromiseSource<HandlingResult>
+    {
+        /// <inheritdoc/>
+        public override bool IsCompleted => true;
+
+
+        /// <inheritdoc/>
+        public override HandlingResult GetResult() => InnerSource;
+
+        /// <inheritdoc/>
+        public override void OnCompleted(Action continuation) => 
+            continuation();
+    }
+}
