+++ Native/Messages/TaskBased/DefaultMulticastMessageHandler.cs
@@ -1,11 +1,10 @@
-﻿using System;
+﻿using Chopsticks.Messages.Abstractions;
+using System;
 using System.Threading;
-using System.Threading.Tasks;
 
-namespace Chopsticks.Messages.TaskBased
+namespace Chopsticks.Messages
 {
-    // TODO :: Update for async abstraction.
-    public static class DefaultTaskMessageHandler<TMessage>
+    public static class DefaultMulticastMessageHandler<TMessage>
     {
         // TODO :: Add an overridable builder for the default handler (supports testing and flexibility).
 
@@ -17,9 +16,9 @@ namespace Chopsticks.Messages.TaskBased
         public static HandlingPromise Handle(TMessage message) =>
             _defaultHandler.Value.Handle(message);
 
-        //public static Task<HandlingResult> HandleAsync(TMessage message,
-        //    CancellationToken token = default) =>
-        //    _defaultHandler.Value.HandleAsync(message, token);
+        public static HandlingAwaitable HandleAsync(TMessage message,
+            CancellationToken token = default) =>
+            _defaultHandler.Value.HandleAsync(message, token);
 
         public static void Register(ISyncMessageHandler<TMessage> handler, 
             params IIntercept<TMessage>[] interceptors)
@@ -27,8 +26,8 @@ namespace Chopsticks.Messages.TaskBased
             if (handler == null)
                 throw new ArgumentNullException(nameof(handler));
 
-            //(_defaultHandler.Value as ITaskMessageHandlerRegistrar<TMessage>)
-              //  .Register(handler, default, interceptors);
+            (_defaultHandler.Value as IMessageHandlerRegistrar<TMessage>)
+                .Register(handler, default, interceptors);
         }
 
         public static void Register(ISyncMessageHandler<TMessage> handler, 
@@ -38,8 +37,8 @@ namespace Chopsticks.Messages.TaskBased
             if (handler == null)
                 throw new ArgumentNullException(nameof(handler));
 
-            //(_defaultHandler.Value as ITaskMessageHandlerRegistrar<TMessage>)
-              //  .Register(handler, settings);
+            (_defaultHandler.Value as IMessageHandlerRegistrar<TMessage>)
+                .Register(handler, settings);
         }
 
         public static void Unregister(ISyncMessageHandler<TMessage> handler)
@@ -47,8 +46,8 @@ namespace Chopsticks.Messages.TaskBased
             if (handler == null)
                 throw new ArgumentNullException(nameof(handler));
 
-            //(_defaultHandler.Value as ITaskMessageHandlerRegistrar<TMessage>)
-              //  .Unregister(handler);
+            (_defaultHandler.Value as IMessageHandlerRegistrar<TMessage>)
+                .Unregister(handler);
         }
     }
 }
