--- Native/Messages/TaskBased/ITaskMessageHandlerRegistrar.cs
@@ -1,8 +1,0 @@
-﻿using Chopsticks.Messages.Abstractions;
-
-namespace Chopsticks.Messages.TaskBased
-{
-    public interface ITaskMessageHandlerRegistrar<TMessage> :
-        IMessageHandlerRegistrar<TMessage>
-    { }
-}
